split array 2d and 3d into their own files
